[OPENMEETINGS-2239] sip room count works better

+++ openmeetings-core/src/main/java/org/apache/openmeetings/core/remote/KRoom.java
@@ -67,7 +67,7 @@ public class KRoom {
 	private final AtomicBoolean recordingStarted = new AtomicBoolean(false);
 	private final AtomicBoolean sharingStarted = new AtomicBoolean(false);
 	private Long recordingId = null;
-	private int sipCount = 0;
+	private long sipCount = 0;
 	private JSONObject recordingUser = new JSONObject();
 	private JSONObject sharingUser = new JSONObject();
 
@@ -249,14 +249,14 @@ public class KRoom {
 		log.debug("Room {} closed", room.getId());
 	}
 
-	public void updateSipCount(final int count) {
+	public void updateSipCount(final long count) {
 		if (count != sipCount) {
 			sipCount = count;
 			processor.getByRoom(room.getId()).forEach(stream -> stream.addSipProcessor(count));
 		}
 	}
 
-	public int getSipCount() {
+	public long getSipCount() {
 		return sipCount;
 	}
 }
