+++ openmeetings-core/src/main/java/org/apache/openmeetings/core/remote/KStream.java
@@ -488,7 +488,7 @@ public class KStream extends AbstractStream implements ISipCallbacks {
 		}
 	}
 
-	void addSipProcessor(int count) {
+	void addSipProcessor(long count) {
 		if (count > 0) {
 			if (sipProcessor.isEmpty()) {
 				try {
